requirements.txt
django
pillow
pytesseract
google-generativeai